pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "libexplorer"
version = "0.1.0"
description = "Discover real-world usage examples of any Python library from GitHub"
requires-python = ">=3.12"
dependencies = [
    "typer[all]>=0.12",
    "httpx>=0.27",
    "pydantic>=2.7",
    "jinja2>=3.1",
    "rich>=13.7",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-httpx>=0.30",
]

[project.scripts]
libexplorer = "libexplorer.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
